re-shape hat tree to 14 bits per level (compiling)

+++ src/Hats.sol
@@ -36,11 +36,11 @@ contract Hats is IHats, ERC1155, HatsIdUtilities {
     struct Hat {
         // 1st storage slot
         address eligibility; // ─┐ can revoke Hat based on ruling | 20
-        uint32 maxSupply; //  │ the max number of identical hats that can exist | 4
-        uint8 lastHatId; // ─┘ indexes how many different hats an admin is holding | 1
+        uint32 maxSupply; //     │ the max number of identical hats that can exist | 4
+        uint16 lastHatId; //    ─┘ indexes how many different hats an admin is holding | 1
         // 2nd slot
         address toggle; // ─┐ controls when Hat is active | 20
-        uint96 config; // ─┘ active status & other settings (see schema below) | 12
+        uint96 config; //  ─┘ active status & other settings (see schema below) | 12
         // 3rd+ slot (optional)
         string details;
         string imageURI;
@@ -237,7 +237,7 @@ contract Hats is IHats, ERC1155, HatsIdUtilities {
     }
 
     function getNextId(uint256 _admin) public view returns (uint256) {
-        uint8 nextHatId = _hats[_admin].lastHatId + 1;
+        uint16 nextHatId = _hats[_admin].lastHatId + 1;
         return buildHatId(_admin, nextHatId);
     }
 
@@ -657,7 +657,7 @@ contract Hats is IHats, ERC1155, HatsIdUtilities {
             address eligibility,
             address toggle,
             string memory imageURI,
-            uint8 lastHatId,
+            uint16 lastHatId,
             bool mutable_,
             bool active
         )
